Refactor classic card presenter study logic into dedicated module
