+++ src/ui/views/classic/cardsPresenter.js
@@ -1,9 +1,6 @@
 import {
   getAssetGallery,
   getHustleControls,
-  getSlideOverNodes,
-  getStudyQueue,
-  getStudyTrackList,
   getUpgradeDockList,
   getUpgradeEmptyNode,
   getUpgradeLaneList,
@@ -11,15 +8,13 @@ import {
   getUpgradeOverview
 } from '../../elements/registry.js';
 import { getAssetState, getState } from '../../../core/state.js';
-import { formatDays, formatHours, formatMoney } from '../../../core/helpers.js';
+import { formatHours, formatMoney } from '../../../core/helpers.js';
 import { describeHustleRequirements, getHustleDailyUsage } from '../../../game/hustles/helpers.js';
 import {
   assetRequirementsMetById,
   describeRequirement,
   formatAssetRequirementLabel,
   getDefinitionRequirements,
-  KNOWLEDGE_TRACKS,
-  getKnowledgeProgress,
   listAssetRequirementDescriptors
 } from '../../../game/requirements.js';
 import {
@@ -67,15 +62,23 @@ import {
   getAssetGroupId,
   getAssetGroupLabel,
   getAssetGroupNote,
-  getUpgradeSnapshot,
-  resolveTrack
+  getUpgradeSnapshot
 } from '../../cards/model.js';
+import { createBadge } from './components/badge.js';
+import { createDefinitionSummary } from './components/definitionSummary.js';
+import { showSlideOver } from './components/slideOver.js';
+import {
+  cacheEducationModels,
+  isStudyDefinition,
+  renderStudySection,
+  updateStudySection,
+  updateStudyTrack
+} from './studySection.js';
 
 const hustleUi = new Map();
 const upgradeUi = new Map();
 const upgradeSections = new Map();
 const upgradeLaneItems = new Map();
-const studyUi = new Map();
 let currentAssetDefinitions = [];
 let currentAssetModels = { groups: [], launchers: [] };
 const assetGroupUi = new Map();
@@ -90,8 +93,6 @@ const upgradeDefinitionLookup = new Map();
 let assetLaunchPanelExpanded = false;
 
 const hustleModelCache = new Map();
-let educationModelCache = null;
-
 function normalizeRegistries(registries = {}) {
   return {
     hustles: Array.isArray(registries?.hustles) ? registries.hustles : [],
@@ -154,7 +155,7 @@ function cacheCardModels(models = {}) {
       hustleModelCache.set(model.id, model);
     }
   });
-  educationModelCache = models?.education ?? null;
+  cacheEducationModels(models?.education);
   cacheAssetModels(models?.assets);
   cacheUpgradeModels(models?.upgrades);
 }
@@ -195,75 +196,6 @@ function findUpgradeModelById(id) {
   return null;
 }
 
-
-function showSlideOver({ eyebrow, title, body }) {
-  const {
-    slideOver,
-    slideOverContent,
-    slideOverEyebrow,
-    slideOverTitle
-  } = getSlideOverNodes() || {};
-  if (!slideOver || !slideOverContent) return;
-  slideOverEyebrow.textContent = eyebrow || '';
-  slideOverTitle.textContent = title || '';
-  slideOverContent.innerHTML = '';
-  if (Array.isArray(body)) {
-    body.forEach(node => slideOverContent.appendChild(node));
-  } else if (body instanceof Node) {
-    slideOverContent.appendChild(body);
-  } else if (typeof body === 'string') {
-    const p = document.createElement('p');
-    p.textContent = body;
-    slideOverContent.appendChild(p);
-  }
-  slideOver.hidden = false;
-  slideOver.focus();
-}
-
-function createBadge(label) {
-  const span = document.createElement('span');
-  span.className = 'badge';
-  span.textContent = label;
-  return span;
-}
-
-function createDefinitionSummary(title, rows = []) {
-  const section = document.createElement('section');
-  const heading = document.createElement('h3');
-  heading.textContent = title;
-  section.appendChild(heading);
-  const list = document.createElement('ul');
-  list.className = 'definition-list';
-  rows.forEach(row => {
-    const item = document.createElement('li');
-    if (row.label) {
-      const label = document.createElement('span');
-      label.textContent = row.label;
-      label.className = 'definition-list__label';
-      item.appendChild(label);
-    }
-    if (row.value) {
-      const value = document.createElement('span');
-      if (row.value instanceof Node) {
-        value.appendChild(row.value);
-      } else {
-        value.textContent = row.value;
-      }
-      value.className = 'definition-list__value';
-      item.appendChild(value);
-    }
-    list.appendChild(item);
-  });
-  section.appendChild(list);
-  return section;
-}
-
-function describeSkillWeight(weight = 0) {
-  if (weight >= 0.75) return 'Signature focus';
-  if (weight >= 0.5) return 'Core boost';
-  if (weight >= 0.3) return 'Supporting practice';
-  return 'Quick primer';
-}
 
 function createAssetDetailHighlights(definition) {
   const entries = Array.isArray(definition.detailEntries)
@@ -3149,305 +3081,12 @@ function renderUpgradeDock() {
   });
 }
 
-function formatStudyCountdown(trackInfo, progress) {
-  if (progress.completed) {
-    return 'Diploma earned';
-  }
-
-  const totalDays = Math.max(0, Number(progress.totalDays ?? trackInfo.days ?? 0));
-  if (!progress.enrolled) {
-    return `${formatDays(totalDays || trackInfo.days)}`;
-  }
-
-  const completedDays = Math.max(0, Math.min(totalDays, Number(progress.daysCompleted) || 0));
-  const remainingDays = Math.max(0, totalDays - completedDays);
-  if (remainingDays === 0) {
-    return 'Graduation tomorrow';
-  }
-  if (remainingDays === 1) {
-    return '1 day remaining';
-  }
-  return `${remainingDays} days remaining`;
-}
-
-function describeStudyMomentum(trackInfo, progress) {
-  if (progress.completed) {
-    return 'Knowledge unlocked for every requirement. Toast your success!';
-  }
-  if (!progress.enrolled) {
-    const tuitionNote = trackInfo.tuition > 0 ? `Pay $${formatMoney(trackInfo.tuition)} upfront and` : 'Just';
-    return `${tuitionNote} we’ll reserve ${formatHours(trackInfo.hoursPerDay)} each day once you enroll.`;
-  }
-  if (progress.studiedToday) {
-    return '✅ Today’s session is logged. Keep the streak cozy until sundown.';
-  }
-  return `Reserve ${formatHours(trackInfo.hoursPerDay)} today to keep momentum humming.`;
-}
-
-function buildStudyBadges(progress) {
-  const badges = [];
-  if (progress.completed) {
-    badges.push(createBadge('Graduated'));
-  } else if (progress.enrolled) {
-    badges.push(createBadge('Enrolled'));
-    badges.push(createBadge(progress.studiedToday ? 'Logged today' : 'Study pending'));
-  } else {
-    badges.push(createBadge('Ready to enroll'));
-  }
-  return badges;
-}
-
-function applyStudyTrackState(track, trackInfo, progress) {
-  track.dataset.active = progress.enrolled ? 'true' : 'false';
-  track.dataset.complete = progress.completed ? 'true' : 'false';
-
-  const countdown = track.querySelector('.study-track__countdown');
-  if (countdown) {
-    countdown.textContent = formatStudyCountdown(trackInfo, progress);
-  }
-
-  const status = track.querySelector('.study-track__status');
-  if (status) {
-    status.innerHTML = '';
-    buildStudyBadges(progress).forEach(badge => status.appendChild(badge));
-  }
-
-  const note = track.querySelector('.study-track__note');
-  if (note) {
-    note.textContent = describeStudyMomentum(trackInfo, progress);
-  }
-
-  const totalDays = Math.max(0, Number(progress.totalDays ?? trackInfo.days ?? 0));
-  const completedDays = progress.completed
-    ? totalDays
-    : Math.max(0, Math.min(totalDays, Number(progress.daysCompleted) || 0));
-  const remainingDays = Math.max(0, totalDays - completedDays);
-  const percent = Math.min(
-    100,
-    Math.max(0, Math.round((totalDays === 0 ? (progress.completed ? 1 : 0) : completedDays / totalDays) * 100))
-  );
-  const fill = track.querySelector('.study-track__progress span');
-  if (fill) {
-    fill.style.width = `${percent}%`;
-    fill.setAttribute('aria-valuenow', String(percent));
-  }
-
-  const progressLabel = track.querySelector('.study-track__progress');
-  if (progressLabel) {
-    progressLabel.setAttribute('aria-label', `${trackInfo.name} progress: ${percent}%`);
-  }
-
-  const remaining = track.querySelector('.study-track__remaining');
-  if (remaining) {
-    const totalLabel = totalDays || trackInfo.days;
-    remaining.textContent = `${completedDays}/${totalLabel} days complete`;
-  }
-
-  const countdownValue = track.querySelector('.study-track__remaining-days');
-  if (countdownValue) {
-    countdownValue.textContent = progress.completed
-      ? 'Course complete'
-      : remainingDays === 1
-        ? '1 day left'
-        : `${remainingDays} days left`;
-  }
-}
-
-function renderStudyTrack(definition) {
-  const state = getState();
-  const trackInfo = resolveTrack(definition);
-  const progress = getKnowledgeProgress(trackInfo.id, state);
-  const track = document.createElement('article');
-  track.className = 'study-track';
-  track.dataset.track = trackInfo.id;
-  track.setAttribute('aria-label', `${trackInfo.name} study track`);
-
-  const header = document.createElement('header');
-  header.className = 'study-track__header';
-
-  const titleGroup = document.createElement('div');
-  titleGroup.className = 'study-track__title-group';
-
-  const title = document.createElement('h3');
-  title.textContent = trackInfo.name;
-  titleGroup.appendChild(title);
-
-  const status = document.createElement('div');
-  status.className = 'study-track__status badges';
-  titleGroup.appendChild(status);
-
-  header.appendChild(titleGroup);
-
-  const countdown = document.createElement('span');
-  countdown.className = 'study-track__countdown';
-  header.appendChild(countdown);
-  track.appendChild(header);
-
-  const summary = document.createElement('p');
-  summary.className = 'study-track__summary';
-  summary.textContent = trackInfo.summary || '';
-  track.appendChild(summary);
-
-  const meta = document.createElement('dl');
-  meta.className = 'study-track__meta';
-  const metaItems = [
-    { label: 'Daily load', value: `${formatHours(trackInfo.hoursPerDay)} / day` },
-    { label: 'Course length', value: formatDays(trackInfo.days) },
-    { label: 'Tuition', value: trackInfo.tuition > 0 ? `$${formatMoney(trackInfo.tuition)}` : 'Free' }
-  ];
-  metaItems.forEach(item => {
-    const dt = document.createElement('dt');
-    dt.textContent = item.label;
-    meta.appendChild(dt);
-    const dd = document.createElement('dd');
-    dd.textContent = item.value;
-    meta.appendChild(dd);
-  });
-  track.appendChild(meta);
-
-  if (trackInfo.skills.length) {
-    const skills = document.createElement('section');
-    skills.className = 'study-track__skills';
-
-    const heading = document.createElement('h4');
-    heading.className = 'study-track__skills-heading';
-    heading.textContent = 'Skill rewards';
-    skills.appendChild(heading);
-
-    if (trackInfo.skillXp > 0) {
-      const xpNote = document.createElement('p');
-      xpNote.className = 'study-track__skills-note';
-      xpNote.textContent = `Graduates collect +${trackInfo.skillXp} XP across these disciplines.`;
-      skills.appendChild(xpNote);
-    }
-
-    const list = document.createElement('ul');
-    list.className = 'study-track__skills-list';
-    trackInfo.skills.forEach(entry => {
-      const item = document.createElement('li');
-      item.className = 'study-track__skills-item';
-      const name = document.createElement('strong');
-      name.textContent = entry.name;
-      item.appendChild(name);
-      const note = document.createElement('span');
-      note.textContent = describeSkillWeight(entry.weight);
-      item.appendChild(note);
-      list.appendChild(item);
-    });
-    skills.appendChild(list);
-    track.appendChild(skills);
-  }
-
-  const progressWrap = document.createElement('div');
-  progressWrap.className = 'study-track__progress-wrap';
-
-  const remaining = document.createElement('span');
-  remaining.className = 'study-track__remaining';
-  progressWrap.appendChild(remaining);
-
-  const bar = document.createElement('div');
-  bar.className = 'study-track__progress';
-  bar.setAttribute('role', 'progressbar');
-  bar.setAttribute('aria-valuemin', '0');
-  bar.setAttribute('aria-valuemax', '100');
-  const fill = document.createElement('span');
-  bar.appendChild(fill);
-  progressWrap.appendChild(bar);
-
-  const remainingDays = document.createElement('span');
-  remainingDays.className = 'study-track__remaining-days';
-  progressWrap.appendChild(remainingDays);
-  track.appendChild(progressWrap);
-
-  const note = document.createElement('p');
-  note.className = 'study-track__note';
-  track.appendChild(note);
-
-  const actions = document.createElement('div');
-  actions.className = 'hustle-card__actions';
-  if (trackInfo.action?.onClick) {
-    const button = document.createElement('button');
-    button.type = 'button';
-    button.className = 'primary';
-    button.textContent = typeof trackInfo.action.label === 'function'
-      ? trackInfo.action.label(state)
-      : trackInfo.action.label || 'Study';
-    button.addEventListener('click', () => trackInfo.action.onClick());
-    actions.appendChild(button);
-  }
-  const details = document.createElement('button');
-  details.type = 'button';
-  details.className = 'ghost';
-  details.textContent = 'Details';
-  details.addEventListener('click', () => openStudyDetails(trackInfo));
-  actions.appendChild(details);
-  track.appendChild(actions);
-
-  applyStudyTrackState(track, trackInfo, progress);
-
-  return { track };
-}
-
-function openStudyDetails(definition) {
-  const body = document.createElement('div');
-  body.className = 'study-detail';
-  if (definition.description) {
-    const intro = document.createElement('p');
-    intro.textContent = definition.description;
-    body.appendChild(intro);
-  }
-  body.appendChild(
-    createDefinitionSummary('Per-level bonuses', (definition.levels || []).map(level => ({
-      label: `Level ${level.level}`,
-      value: level.name
-    })))
-  );
-  showSlideOver({ eyebrow: 'Study track', title: definition.name, body });
-}
-
-function renderEducation(definitions, educationModels) {
-  const list = getStudyTrackList();
-  if (!list) return;
-  list.innerHTML = '';
-  studyUi.clear();
-  definitions.forEach(def => {
-    const { track } = renderStudyTrack(def);
-    list.appendChild(track);
-    studyUi.set(resolveTrack(def).id, { track });
-  });
-  renderStudyQueue(educationModels);
-}
-
-function renderStudyQueue(educationModels) {
-  const { list: queue, eta: queueEta, cap: capNode } = getStudyQueue() || {};
-  if (!queue) return;
-  queue.innerHTML = '';
-  const queueModel = educationModels?.queue;
-  (queueModel?.entries ?? []).forEach(entry => {
-    const item = document.createElement('li');
-    item.textContent = `${entry.name} • ${formatHours(entry.hoursPerDay)} per day`;
-    queue.appendChild(item);
-  });
-  if (!queue.childElementCount) {
-    const empty = document.createElement('li');
-    empty.textContent = 'No study queued today.';
-    queue.appendChild(empty);
-  }
-  if (queueEta) {
-    queueEta.textContent = queueModel?.totalLabel || '';
-  }
-
-  if (capNode) {
-    capNode.textContent = queueModel?.capLabel || '';
-  }
-}
-
 function renderClassicCollections(registries, models) {
   const { hustles = [], education = [], assets = [], upgrades = [] } = registries;
   renderHustles(hustles, models?.hustles ?? []);
   renderAssets(assets, models?.assets ?? currentAssetModels);
   renderUpgrades(upgrades, models?.upgrades);
-  renderEducation(education, models?.education ?? educationModelCache);
+  renderStudySection(education, models?.education);
 }
 
 export function renderAll({ registries = {}, models = {} } = {}) {
@@ -3475,11 +3114,8 @@ export function updateCard(definition) {
     emitUIEvent('upgrades:state-updated');
     return;
   }
-  if (definition.tag?.type === 'study' || KNOWLEDGE_TRACKS[definition.id]) {
-    const trackInfo = resolveTrack(definition);
-    if (studyUi.has(trackInfo.id)) {
-      updateStudyTrack(definition);
-    }
+  if (isStudyDefinition(definition)) {
+    updateStudyTrack(definition);
   }
 }
 
@@ -3492,12 +3128,7 @@ function updateClassicCollections(registries, models) {
   });
   updateAssets(assets, models?.assets ?? currentAssetModels);
   updateUpgrades(upgrades, models?.upgrades);
-  education.forEach(def => {
-    if (def.tag?.type === 'study' || KNOWLEDGE_TRACKS[def.id]) {
-      updateStudyTrack(def);
-    }
-  });
-  renderStudyQueue(models?.education ?? educationModelCache);
+  updateStudySection(education, models?.education);
   renderUpgradeDock();
   refreshUpgradeSections();
   emitUIEvent('upgrades:state-updated');
@@ -3507,15 +3138,6 @@ export function update({ registries = {}, models = {} } = {}) {
   const normalized = normalizeRegistries(registries);
   cacheCardModels(models);
   updateClassicCollections(normalized, models);
-}
-
-function updateStudyTrack(definition) {
-  const info = resolveTrack(definition);
-  const ui = studyUi.get(info.id);
-  if (!ui) return;
-  const state = getState();
-  const progress = getKnowledgeProgress(info.id, state);
-  applyStudyTrackState(ui.track, info, progress);
 }
 
 const classicCardsPresenter = {
