+++ src/ui/views/classic/components/badge.js
@@ -1,0 +1,8 @@
+export function createBadge(label) {
+  const span = document.createElement('span');
+  span.className = 'badge';
+  span.textContent = label;
+  return span;
+}
+
+export default createBadge;
