+++ src/ui/views/classic/components/definitionSummary.js
@@ -1,0 +1,32 @@
+export function createDefinitionSummary(title, rows = []) {
+  const section = document.createElement('section');
+  const heading = document.createElement('h3');
+  heading.textContent = title;
+  section.appendChild(heading);
+  const list = document.createElement('ul');
+  list.className = 'definition-list';
+  rows.forEach(row => {
+    const item = document.createElement('li');
+    if (row.label) {
+      const label = document.createElement('span');
+      label.textContent = row.label;
+      label.className = 'definition-list__label';
+      item.appendChild(label);
+    }
+    if (row.value) {
+      const value = document.createElement('span');
+      if (row.value instanceof Node) {
+        value.appendChild(row.value);
+      } else {
+        value.textContent = row.value;
+      }
+      value.className = 'definition-list__value';
+      item.appendChild(value);
+    }
+    list.appendChild(item);
+  });
+  section.appendChild(list);
+  return section;
+}
+
+export default createDefinitionSummary;
